% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{get_time_series}
\alias{get_time_series}
\title{Retrieve data during a time series}
\usage{
get_time_series(data, station, start = NULL, end = NULL)
}
\arguments{
\item{data}{data frame with NCEI data}

\item{station}{station name}

\item{start}{optional argument for start date}

\item{end}{optional argument for end date}
}
\value{
A data frame
}
\description{
Get weather data for each day, unless given specific time series, for a given station.
}
