% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{get_time_series}
\alias{get_time_series}
\title{Retrieve data during a time series}
\usage{
get_time_series(data, station, start = NULL, end = NULL)
}
\arguments{
\item{data}{data frame with NCEI data}

\item{station}{station name}

\item{start}{optional argument for start date, a Date object}

\item{end}{optional argument for end date, a Date object}
}
\value{
A data frame
}
\description{
Get weather data for each day, unless given specific time series, for a given station.
}
\examples{
# Using built in dataset
data <- daily_weather::data
start <- as.Date("20010101","\%Y\%m\%d")
end <- as.Date("20231231","\%Y\%m\%d")
ex1 <- get_time_series(data, "NC_Asheville_13_S", start = start)
ex2 <- get_time_series(data, "NC_Asheville_13_S", end = end)
print(head(ex1$LST_DATE))
print(tail(ex2$LST_DATE))

}
